refactor: return extras.RulesExtra from GetRulesExtra
