+++ params/extras/rules_extra.go
@@ -24,3 +24,18 @@ type RulesExtra struct {
 	// that are enabled for this rule set.
 	AccepterPrecompiles map[common.Address]precompileconfig.Accepter
 }
+
+func (r *RulesExtra) PredicatersExist() bool {
+	return len(r.Predicaters) > 0
+}
+
+func (r *RulesExtra) PredicaterExists(addr common.Address) bool {
+	_, ok := r.Predicaters[addr]
+	return ok
+}
+
+// IsPrecompileEnabled returns true if the precompile at [addr] is enabled for this rule set.
+func (r *RulesExtra) IsPrecompileEnabled(addr common.Address) bool {
+	_, ok := r.Precompiles[addr]
+	return ok
+}
