+++ plugin/evm/import_tx.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/ava-labs/coreth/core/state"
 	"github.com/ava-labs/coreth/params"
+	"github.com/ava-labs/coreth/params/extras"
 	"github.com/holiman/uint256"
 
 	"github.com/ava-labs/avalanchego/chains/atomic"
@@ -62,7 +63,7 @@ func (utx *UnsignedImportTx) InputUTXOs() set.Set[ids.ID] {
 // Verify this transaction is well-formed
 func (utx *UnsignedImportTx) Verify(
 	ctx *snow.Context,
-	rules params.RulesExtra,
+	rules extras.RulesExtra,
 ) error {
 	switch {
 	case utx == nil:
@@ -181,7 +182,7 @@ func (utx *UnsignedImportTx) SemanticVerify(
 	stx *Tx,
 	parent *Block,
 	baseFee *big.Int,
-	rules params.RulesExtra,
+	rules extras.RulesExtra,
 ) error {
 	if err := utx.Verify(vm.ctx, rules); err != nil {
 		return err
