+++ plugin/evm/tx_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/ava-labs/libevm/common"
 
 	"github.com/ava-labs/coreth/params"
+	"github.com/ava-labs/coreth/params/extras"
 
 	"github.com/ava-labs/avalanchego/chains/atomic"
 	"github.com/ava-labs/avalanchego/ids"
@@ -60,7 +61,7 @@ func TestCalculateDynamicFee(t *testing.T) {
 type atomicTxVerifyTest struct {
 	ctx         *snow.Context
 	generate    func(t *testing.T) UnsignedAtomicTx
-	rules       params.RulesExtra
+	rules       extras.RulesExtra
 	expectedErr string
 }
